pyproject.toml
[project]
name = "ria"
version = "0.1.0"
description = ""
authors = [
    {name = "Prabeg Shakya",email = "shakyaprabeg@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard] (>=0.121.2,<0.122.0)",
    "google-genai (>=1.50.1,<2.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "pymupdf (>=1.26.6,<2.0.0)",
    "rq (>=2.6.0,<3.0.0)",
    "aiofiles (>=25.1.0,<26.0.0)",
    "psycopg[binary,pool] (>=3.2.13,<4.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "boto3 (>=1.40.76,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
black = "^25.11.0"

